+++ apps/dashboard/app/__init__.py
@@ -1,1 +1,3 @@
-__version__ = "0.0.2"
+__version__ = "0.1.2"
+__author__ = "Adam Stone"
+__github__ = "https://github.com/ast0ne1"